make individuals serializable & refactor 'save_current_results'
