+++ fedot/core/optimisers/opt_history_objects/opt_history.py
@@ -16,7 +16,7 @@ from fedot.core.optimisers.objective import Objective
 from fedot.core.optimisers.utils.population_utils import get_metric_position
 from fedot.core.pipelines.template import PipelineTemplate
 from fedot.core.repository.quality_metrics_repository import QualityMetricsEnum
-from fedot.core.serializers import Serializer
+from fedot.core.serializers.serializer import default_load, default_save
 from fedot.core.utils import default_fedot_data_dir
 from fedot.core.visualisation.opt_viz import OptHistoryVisualizer
 
@@ -81,39 +81,20 @@ class OptHistory:
             last_gen_id = len(self.individuals) - 1
             last_gen = self.individuals[last_gen_id]
             last_gen_history = self.historical_fitness[last_gen_id]
-            adapter = PipelineAdapter()
             for individual, ind_fitness in zip(last_gen, last_gen_history):
-                ind_path = Path(save_dir, str(last_gen_id), str(individual.uid))
-                additional_info = \
-                    {'fitness_name': self._objective.metric_names,
-                     'fitness_value': ind_fitness}
-                PipelineTemplate(adapter.restore(individual)).\
-                    export_pipeline(path=ind_path, additional_info=additional_info, datetime_in_path=False)
+                ind_path = Path(save_dir, str(last_gen_id))
+                if not os.path.isdir(ind_path):
+                    os.makedirs(ind_path)
+                individual.save(json_file_path=Path(ind_path, f'{str(individual.uid)}.json'))
         except Exception as ex:
             self._log.exception(ex)
 
     def save(self, json_file_path: Union[str, os.PathLike] = None) -> Optional[str]:
-        if json_file_path is None:
-            return json.dumps(self, indent=4, cls=Serializer)
-        with open(json_file_path, mode='w') as json_file:
-            json.dump(self, json_file, indent=4, cls=Serializer)
+        return default_save(obj=self, json_file_path=json_file_path)
 
     @staticmethod
     def load(json_str_or_file_path: Union[str, os.PathLike] = None) -> 'OptHistory':
-        def load_as_file_path():
-            with open(json_str_or_file_path, mode='r') as json_file:
-                return json.load(json_file, cls=Serializer)
-
-        def load_as_json_str():
-            return json.loads(json_str_or_file_path, cls=Serializer)
-
-        if isinstance(json_str_or_file_path, os.PathLike):
-            return load_as_file_path()
-
-        try:
-            return load_as_json_str()
-        except json.JSONDecodeError:
-            return load_as_file_path()
+        return default_load(json_str_or_file_path)
 
     @staticmethod
     def clean_results(dir_path: Optional[str] = None):
